backend/requirements.txt
fastapi
python-multipart
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv
langgraph
langchain-core
httpx
tenacity
rich
pypdf
